refactor: 디렉터리 구조 및 파일 네이밍 컨벤션 변경
