--- packages/cli/src/api/index.ts
@@ -1,4 +1,0 @@
-export * from './getLeetCodeDailyChallenge';
-export * from './getLeetCodeQuestionBySlug';
-export * from './getProgrammersProblem';
-export * from './getLeetCodeSlugById';
