+++ packages/cli/src/api/programmers/index.ts
@@ -1,0 +1,1 @@
+export { getProgrammersProblemById } from './problem-by-id';
