+++ packages/cli/src/lib/generate/leetcode.ts
@@ -1,6 +1,10 @@
 import path from 'node:path';
 
-import { getLeetCodeQuestionBySlug, getLeetCodeSlugById } from '@/api';
+import {
+  getLeetCodeDailyChallenge,
+  getLeetCodeQuestionBySlug,
+  getLeetCodeSlugById,
+} from '@/api/leetcode';
 import { EXTNAME } from '@/constants';
 import { createSolutionTemplate, createTestTemplate } from '@/lib/template';
 
@@ -28,6 +32,28 @@ export async function generateLeetCode(outputDir: string, id: string) {
     ensureWriteFile(
       path.join(datePath, `${fileName}${EXTNAME.TYPESCRIPT_TEST}`),
       createTestTemplate(CodingSite.LeetCode, dateString, id),
+    ),
+  ]);
+}
+
+export async function generateLeetCodeDailyChallenge(outputDir: string) {
+  const {
+    date,
+    question: { titleSlug },
+  } = await getLeetCodeDailyChallenge();
+  const { id, title, codeSnippet, testCases } = await getLeetCodeQuestionBySlug(titleSlug);
+  const datePath = getDatePath(new Date(date), outputDir);
+
+  const fileName = kebabcase(CodingSite.LeetCode, id);
+
+  await Promise.all([
+    ensureWriteFile(
+      path.join(datePath, `${fileName}${EXTNAME.TYPESCRIPT}`),
+      createSolutionTemplate(CodingSite.LeetCode, id, title, titleSlug, codeSnippet),
+    ),
+    ensureWriteFile(
+      path.join(datePath, `${fileName}${EXTNAME.TYPESCRIPT_TEST}`),
+      createTestTemplate(CodingSite.LeetCode, date, id, testCases),
     ),
   ]);
 }
