+++ packages/cli/src/lib/generate/readme.ts
