+++ packages/cli/src/lib/parse/index.ts
@@ -1,2 +1,2 @@
-export * from './parseLeetCode';
-export * from './parseProgrammers';
+export { parseLeetCode } from './leetcode';
+export { parseProgrammers } from './programmers';
