+++ packages/cli/src/lib/prompt/generate-type.ts
