+++ packages/cli/src/lib/template/readme.ts
