+++ packages/cli/src/lib/template/solution.ts
