+++ packages/cli/src/lib/template/test.ts
